requirements.txt
pandas
pillow
python-dotenv
openpyxl
google-generativeai
torch
xformers
transformers
torchvision
einops
tensorflow
matplotlib
timm
sentencepiece
accelerate
apex
modelscope
tiktoken
transformers_stream_generator
tqdm
